requirements.txt
hurry.filesize==0.9
flask==0.12.2
pymongo==3.6.0
numpy==1.6.2
gunicorn==20.0.4
